Added first implementation of a asynchronous dispatcher.
A Command together with it's interceptor chain will be dispatched on a
new thread. This implies that for now, the validation, logging and
possible any custom interceptor is executed asynchronously.

I'm not convinced this is ideal, maybe it would be nice if validation is
done synchronously before dispatching the command to a new thread. On
the other hand, if validation is done before dispatching to a new thread
it could be that other commands changed the state in such a way that the
command should not have been executed anymore.