first version of the eventstore support
An in-memory version of the event store which is capable of recreating
aggregate root entities based on events.

In order for this to work the aggregate root must be annotated to
indicate how the events should be handled. With this information, the
object constructor is able to send the events to an aggregate root in
the order that they got persisted.