Preparing the EventBus to be able to call aggregates.
The EventHandlerInvoker is refactored so that it's an interface allowing
for the event sourcing module to create a different implementation which
can first create an or load an aggregate before calling the event
handler on it.

The EventBus used to create EventHandlerInvoker objects when subscribing
an object to the event bus. This code is extracted to a factory with
interface so that an alternative implementation can be supplied.

What remains now is a mechanism to configure the eventbus to use a
certain implementation of the factory and EventHandlerInvoker.