Added first version of snapshot support
An AggregateRoot object should be able to store snapshots and when
recreating an AggregateRoot we should check if a snapshot exists and if
so, use it to reconstruct the object.

For now we only have an in memory version of the snapshot repository
which is probably not working in a multi threaded environment.

Introduced the notion of a revision which is stored in a snapshot, and
which should be stored in the events as well. Don't know yet how we will
generate this revision though because it should be garantueed to be
incrementing.