Changed signature of the publishing API.
Inspireded by a pull request from @krishofmans.

The EventBus should be responsible for creating a Future object when
calling the publish method. this implies that the client should not have
to bother to create a FutureEventCallback.

This implies that when calling `eventbus.publish(command)` without doing
anything with the returned future object, potential exceptions could be swallowed.