Split up in commands and events first try
We need to split the dispatchers and pull out some of the code of it

A command should have at most one handler, while an event can have
multiple, however, we do not expect an event handler to have a return
type, which is the contrary for a command.

So the dispatchers should be different for both usecases.