Further splitting for commands and events.
Created a Command and Event dispatcher to separate those concerns.

A command is supposed to have only one handler while events are capable
of invoking multiple handlers.

If for a command we find multiple handlers, we throw an exception since
we don't expect multiple.

For events we do not expect any return from the handlers, so we don't do
anyting with them.