Introduction of a UnitOfWork to fix issue #1..
The unit of work gets created when dispatching a command and cleared when
the dispatching is done.

If a CommandHandler uses the EventPublisher to publish events and a unit
of work exists, the events will be delayed until the unit of work gets
committed, which is when the CommandHandler finishes.

Done some cleanup and removed the dispatchers to put that logic back in
the eventbus itself. allowed me to cleanup some duplicate code because
the EventDispatcher and CommandDispatcher where doing more or less the
same thing: Create an interceptorChain with the invokers, and then
starting the chain.

The difference between command and event dispatchers is only that for a
command dispatcher we need to wrap the call using a CallbackTemplate so
that the callback is invoked appropriately.