Interceptors now have access to the unitOfWork.
The unitOfWork which is created when dispatching a command is now passed
to every interceptor. This is usefull to pass information from the
thread that dispatches a command to the thread which is handling the
command.

The demo application is extended to showcase that we can pass the
SecurityContext from the http thread to the command handler thread.